ENH: run.py replaces all scripts.
Previously, the user was running a specific script for each algorithm to run (Low-Rank decomposition, Unbiased Atlas Building, Non-greedy Low-Rank Atlas creation). This was generating duplicate code in this project, and was likely to generate duplicate code for external applications using the pyLAR library (such as pyLAR-SlicerExtension). To simplify this problem, there is now only one script available to the user to run one of the three different algorithm. The user specifies whic algorithm to run with a flag in the command line. The majority of the code has been moved inside the pyLAR module and is directly usable from external project.