ENH: Users can specify in config file the number of processors to use
A new option has been added to the configuration file to allow users to speficy how many cores and processors they want to use while running pyLAR. This option is available when running "Non greedy low rank atlas" (nglra) and "Unbias atlas creation" (uab) to limit the number of processes ('ANTS', 'BRAINSFit',...) run in parallel. If this option is not specified, the default value is set to the number of cores available on the computer.