BUG: Windows compatibility
In command lines containing multiple commands, on Unix systems, commands must be separated by a semi-colon (;).
On Windows, commands must be separated by one ampersand (&).
Additionally, the log handles are closed upon exit of the script. Leave them open on Windows forbids to run
a second script in a row with the same log file as the file is still open.