BUG: File names listed in input list file can now have a space
Previously, a backslash had to be manually added to the path of the
files that contained a space character or an error would occur (the
file was not found). Since input files paths are saves in a text file
in the output directory, the output text file will contain the modified,
i.e with a backlash, file name. Therefore both cases, with
and without a backslash need to be handled.